back-end/requirements.txt
asgiref==3.4.1
atomicwrites==1.3.0
attrs==19.1.0
certifi==2019.3.9
chardet==3.0.4
coverage==4.5.3
Django==3.2.6
django-cors-headers==3.8.0
djangorestframework==3.12.4
Faker==1.0.5
idna==2.8
more-itertools==7.0.0
pluggy==0.9.0
psycopg2==2.9.1
py==1.8.0
pytest==4.4.1
pytest-cov==2.6.1
python-dateutil==2.8.0
pytz==2021.1
requests==2.21.0
requests-mock==1.5.2
six==1.12.0
sqlparse==0.4.1
text-unidecode==1.2
urllib3==1.24.2
